+++ constants/address.ts
@@ -1,10 +1,6 @@
 import { Address } from 'wagmi';
 
-export const KWENTA_REFERRAL_ADDRESS = '0x08e30BFEE9B73c18F9770288DDd13203A4887460';
-
 export const ENS_REVERSE_LOOKUP = '0x3671aE578E63FdF66ad4F3E12CC0c0d71Ac7510C';
-
-export const SYNTH_SWAP_OPTIMISM_ADDRESS = '0x6d6273f52b0C8eaB388141393c1e8cfDB3311De6';
 
 export const CROSS_MARGIN_ACCOUNT_FACTORY: Record<string, string> = {
 	420: '0x9320170B37eDEb4f41cb6E5A8F82B984aD9c44eE',
