--- constants/exchange.ts
@@ -1,1 +1,0 @@
-export const ATOMIC_EXCHANGE_SLIPPAGE = '0.01';
