+++ constants/funding.ts
@@ -1,0 +1,8 @@
+import { FundingRatePeriods } from 'state/futures/types';
+
+export const FUNDING_RATE_PERIODS: FundingRatePeriods = {
+	1: '1H',
+	8: '8H',
+	24: '24H',
+	8766: '1Y',
+};
