--- queries/coingecko/constants.ts
@@ -1,1 +1,0 @@
-export const CG_BASE_API_URL = 'https://api.coingecko.com/api/v3';
