+++ sdk/constants/transactions.ts
@@ -1,6 +1,6 @@
 // Copied over from: https://github.com/Synthetixio/js-monorepo
 
-import { TransactionEventCode } from 'sdk/types/transactions';
+import { TransactionEventCode } from '../types/transactions';
 
 export const TRANSACTION_EVENTS: TransactionEventCode[] = [
 	'txSent',
@@ -12,3 +12,7 @@ export const TRANSACTION_EVENTS: TransactionEventCode[] = [
 export const TRANSACTION_EVENTS_MAP = Object.fromEntries(
 	TRANSACTION_EVENTS.map((event) => [event, event])
 );
+
+export const DEFAULT_GAS_BUFFER = 5000;
+
+export const GWEI_DECIMALS = 9;
