+++ sections/dashboard/Overview/Overview.tsx
@@ -9,9 +9,9 @@ import FuturesIcon from 'components/Nav/FuturesIcon';
 import { TabPanel } from 'components/Tab';
 import Search from 'components/Table/Search';
 import * as Text from 'components/Text';
-import { ETH_ADDRESS, ETH_COINGECKO_ADDRESS } from 'constants/currency';
 import Connector from 'containers/Connector';
 import { FuturesAccountTypes } from 'queries/futures/types';
+import { ETH_ADDRESS, ETH_COINGECKO_ADDRESS } from 'sdk/constants/exchange';
 import { SynthSymbol } from 'sdk/data/synths';
 import { selectBalances } from 'state/balances/selectors';
 import { sdk } from 'state/config';
