+++ state/app/reducer.ts
@@ -4,10 +4,11 @@ import { notifyError } from 'components/ErrorView/ErrorNotifier';
 import { TransactionStatus } from 'sdk/types/common';
 import { FuturesMarketKey } from 'sdk/types/futures';
 import { OperationalStatus } from 'sdk/types/system';
+import { GasPrice } from 'sdk/types/transactions';
 import { isUserDeniedError } from 'utils/formatters/error';
 
 import { checkSynthetixStatus, fetchKwentaStatus } from './actions';
-import { AppState, FuturesPositionModalType, GasPrice, ModalType, Transaction } from './types';
+import { AppState, FuturesPositionModalType, ModalType, Transaction } from './types';
 
 export const APP_INITIAL_STATE: AppState = {
 	showModal: undefined,
