+++ utils/formatters/number.ts
@@ -200,15 +200,6 @@ export const multiplyDecimal = (x: BigNumber, y: BigNumber) => {
 	return x.mul(y).div(UNIT_BIG_NUM);
 };
 
-export const weiFromWei = (weiAmount: WeiSource) => {
-	if (weiAmount instanceof Wei) {
-		const precisionDiff = 18 - weiAmount.p;
-		return wei(weiAmount, 18, true).div(10 ** precisionDiff);
-	} else {
-		return wei(weiAmount, 18, true);
-	}
-};
-
 export const suggestedDecimals = (value: WeiSource) => {
 	value = wei(value).abs().toNumber();
 	if (value >= 100000) return 0;
