+++ utils/network.ts
@@ -1,15 +1,7 @@
-import Wei, { wei } from '@synthetixio/wei';
 import { ethers, providers } from 'ethers';
 
-import { DEFAULT_GAS_BUFFER } from 'constants/defaults';
-import {
-	GWEI_DECIMALS,
-	GasLimitEstimate,
-	SUPPORTED_NETWORKS,
-	BLAST_NETWORK_LOOKUP,
-} from 'constants/network';
+import { SUPPORTED_NETWORKS, BLAST_NETWORK_LOOKUP } from 'constants/network';
 import { NetworkId } from 'sdk/types/common';
-import { GasPrice } from 'state/app/types';
 
 export const staticMainnetProvider = new ethers.providers.InfuraProvider();
 
@@ -55,27 +47,3 @@ export const getDefaultProvider = (networkId: NetworkId) => {
 	}
 	return ethersProvider;
 };
-
-export const getTotalGasPrice = (gasPriceObj?: GasPrice | null) => {
-	if (!gasPriceObj) return wei(0);
-	const { gasPrice, baseFeePerGas, maxPriorityFeePerGas } = gasPriceObj;
-	if (gasPrice) {
-		return wei(gasPrice, GWEI_DECIMALS);
-	}
-	return wei(baseFeePerGas || 0, GWEI_DECIMALS).add(wei(maxPriorityFeePerGas || 0, GWEI_DECIMALS));
-};
-
-export const getTransactionPrice = (
-	gasPrice: GasPrice | null,
-	gasLimit: GasLimitEstimate,
-	ethPrice: Wei | null,
-	optimismLayerOneFee: Wei | null
-) => {
-	if (!gasPrice || !gasLimit || !ethPrice) return null;
-	const totalGasPrice = getTotalGasPrice(gasPrice);
-	const gasPriceCost = totalGasPrice.mul(wei(gasLimit, GWEI_DECIMALS)).mul(ethPrice);
-	const l1Cost = ethPrice.mul(optimismLayerOneFee || 0);
-	return gasPriceCost.add(l1Cost);
-};
-
-export const normalizeGasLimit = (gasLimit: number) => gasLimit + DEFAULT_GAS_BUFFER;
